Make thrift transport configurable
Thrift transport can be specified on a per-connection basis by providing the
fully qualified name of an ITransportFactory implementation. This can be
set using a new 'transportFactory' querystring parameter on the connection
url, or directly on a CassandraDatasource using the setTransportFactory
method. Specifying the factory is optional with the default being to use
TFramedTransport.